+++ src/main/java/com/github/syr0ws/craftventory/api/inventory/CraftVentory.java
@@ -1,6 +1,7 @@
 package com.github.syr0ws.craftventory.api.inventory;
 
 import com.github.syr0ws.craftventory.api.InventoryService;
+import com.github.syr0ws.craftventory.api.inventory.data.DataStore;
 import com.github.syr0ws.craftventory.api.inventory.hook.HookManager;
 import com.github.syr0ws.craftventory.api.inventory.pagination.PaginationManager;
 import com.github.syr0ws.craftventory.api.transform.InventoryProvider;
@@ -113,9 +114,9 @@ public interface CraftVentory {
     /**
      * Retrieves the local storage for the inventory.
      *
-     * @return The {@link InventoryStorage} instance associated with the inventory.
+     * @return The {@link DataStore} instance associated with the inventory.
      */
-    InventoryStorage getLocalStorage();
+    DataStore getLocalStore();
 
     /**
      * Retrieves the service managing the inventory.
