+++ src/main/java/com/github/syr0ws/craftventory/internal/inventory/SimpleCraftVentory.java
@@ -10,7 +10,7 @@ import com.github.syr0ws.craftventory.api.inventory.event.CraftVentoryBeforeOpen
 import com.github.syr0ws.craftventory.api.inventory.exception.InventoryException;
 import com.github.syr0ws.craftventory.api.inventory.hook.HookManager;
 import com.github.syr0ws.craftventory.api.inventory.item.InventoryItem;
-import com.github.syr0ws.craftventory.api.inventory.InventoryStorage;
+import com.github.syr0ws.craftventory.api.inventory.data.DataStore;
 import com.github.syr0ws.craftventory.api.inventory.pagination.PaginationManager;
 import com.github.syr0ws.craftventory.api.transform.InventoryProvider;
 import com.github.syr0ws.craftventory.api.util.Context;
@@ -19,6 +19,7 @@ import com.github.syr0ws.craftventory.common.transform.dto.InventoryTypeDto;
 import com.github.syr0ws.craftventory.common.transform.dto.TitleDto;
 import com.github.syr0ws.craftventory.common.transform.provider.ProviderNameEnum;
 import com.github.syr0ws.craftventory.common.util.CommonContextKey;
+import com.github.syr0ws.craftventory.internal.inventory.data.SimpleDataStore;
 import com.github.syr0ws.craftventory.internal.inventory.hook.SimpleHookManager;
 import com.github.syr0ws.craftventory.internal.inventory.item.SimpleInventoryItem;
 import com.github.syr0ws.craftventory.internal.inventory.pagination.SimplePaginationManager;
@@ -32,7 +33,7 @@ public class SimpleCraftVentory implements CraftVentory {
     private final InventoryProvider provider;
     private final InventoryContent content;
     private final InventoryService service;
-    private final InventoryStorage model;
+    private final DataStore model;
     private final SimplePaginationManager paginationManager;
     private final HookManager hookManager;
     private final InventoryViewer viewer;
@@ -57,7 +58,7 @@ public class SimpleCraftVentory implements CraftVentory {
         this.service = service;
         this.viewer = viewer;
         this.content = new SimpleInventoryContent(this);
-        this.model = new SimpleInventoryStorage();
+        this.model = new SimpleDataStore();
         this.paginationManager = new SimplePaginationManager();
         this.hookManager = new SimpleHookManager();
     }
@@ -148,7 +149,7 @@ public class SimpleCraftVentory implements CraftVentory {
     }
 
     @Override
-    public InventoryStorage getLocalStorage() {
+    public DataStore getLocalStore() {
         return this.model;
     }
