+++ src/main/java/com/github/syr0ws/craftventory/internal/inventory/SimpleInventoryViewManager.java
@@ -1,8 +1,9 @@
 package com.github.syr0ws.craftventory.internal.inventory;
 
 import com.github.syr0ws.craftventory.api.inventory.CraftVentory;
-import com.github.syr0ws.craftventory.api.inventory.InventoryStorage;
+import com.github.syr0ws.craftventory.api.inventory.data.DataStore;
 import com.github.syr0ws.craftventory.api.inventory.InventoryViewManager;
+import com.github.syr0ws.craftventory.internal.inventory.data.SimpleDataStore;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,7 +11,7 @@ import java.util.Optional;
 
 public class SimpleInventoryViewManager implements InventoryViewManager {
 
-    private final InventoryStorage storage = new SimpleInventoryStorage();
+    private final DataStore storage = new SimpleDataStore();
     private final List<CraftVentory> history = new ArrayList<>();
 
     private int index = -1;
@@ -195,7 +196,7 @@ public class SimpleInventoryViewManager implements InventoryViewManager {
     }
 
     @Override
-    public InventoryStorage getSharedStorage() {
+    public DataStore getSharedStore() {
         return this.storage;
     }
